+++ src/main/java/com/todorian/todo/application/controller/TodoController.java
@@ -1,0 +1,14 @@
+package com.todorian.todo.application.controller;
+
+import com.todorian.todo.application.service.TodoService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api")
+public class TodoController {
+    private final TodoService todoService;
+
+}
