+++ src/main/java/com/todorian/todo/domain/repository/TodoRepository.java
@@ -1,0 +1,7 @@
+package com.todorian.todo.domain.repository;
+
+import com.todorian.todo.domain.model.Todo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TodoRepository extends JpaRepository<Todo, Long> {
+}
