+++ src/test/java/com/todorian/todo/service/TodoServiceTest.java
@@ -1,7 +1,8 @@
 package com.todorian.todo.service;
 
-import com.todorian.todo.domain.Todo;
-import com.todorian.todo.dto.TodoCreateRequestDTO;
+import com.todorian.todo.application.service.TodoService;
+import com.todorian.todo.domain.model.Todo;
+import com.todorian.todo.application.dto.TodoCreateRequestDTO;
 import jakarta.persistence.EntityManager;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.Assertions;
@@ -60,7 +61,7 @@ class TodoServiceTest {
     @ParameterizedTest
     @ValueSource(longs = {1})
     void findTodoById(Long todoId) {
-        Todo foundTodo = todoService.findTodoById(todoId);
+        Todo foundTodo = todoService.findTodoById(todoId).orElseThrow();
 
         Assertions.assertNotNull(foundTodo);
         Assertions.assertEquals(todoId, foundTodo.getTodoId());
